feat: Adicionando as classes DTO´s da Categoria, inserção da QueryMethod na interface LivroRepository visando evitar a violação de integridade referencial e CategoriaDeletionException com o status 409(conflict)

+++ src/main/java/com/devlib/devlib/services/CategoriaService.java
@@ -1,14 +1,16 @@
 package com.devlib.devlib.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.devlib.devlib.DTO.CategoriaInsertDTO;
+import com.devlib.devlib.DTO.CategoriaUpdateDTO;
 import com.devlib.devlib.entites.Categoria;
 import com.devlib.devlib.repositories.CategoriaRepository;
-import com.devlib.devlib.services.exceptions.BadRequestException;
+import com.devlib.devlib.repositories.LivroRepository;
+import com.devlib.devlib.services.exceptions.CategoriaDeletionException;
 import com.devlib.devlib.services.exceptions.CategoriaNotFoundException;
 
 @Service
@@ -17,40 +19,36 @@ public class CategoriaService {
 	@Autowired
 	private CategoriaRepository repository;
 	
+	@Autowired
+	private LivroRepository livroRepository;
+	
 	public List<Categoria> findAll(){
 		return repository.findAll();
 	}
 	public Categoria findById(Long id) {
-		validationCategoriaId(id);
-		Optional<Categoria> categoria = repository.findById(id);
-		return categoria.get();
+		Categoria categoria = repository.findById(id)
+				.orElseThrow(() -> new CategoriaNotFoundException(id));
+		return categoria;
 	}
-	public Categoria insert(Categoria categoria) {
-		badRequestId(categoria);
+	public Categoria insert(CategoriaInsertDTO categoriaDTO) {
+		Categoria categoria = new Categoria();
+		categoria.setTitulo(categoriaDTO.getTitulo());
+		categoria.setDescricao(categoriaDTO.getDescricao());
 		return repository.save(categoria);
 	}
-	public Categoria update(Long id, Categoria entity) {
-		validationCategoriaId(id);
-		Categoria categoria = repository.getReferenceById(id);
+	public Categoria update(Long id, CategoriaUpdateDTO entity) {
+		Categoria categoria = repository.findById(id)
+				.orElseThrow(() -> new CategoriaNotFoundException(id));
 		updateData(categoria, entity);
 		return repository.save(categoria);
 	}
-	public void updateData(Categoria categoria, Categoria entity) {
+	public void updateData(Categoria categoria, CategoriaUpdateDTO entity) {
 		categoria.setTitulo(entity.getTitulo());
 		categoria.setDescricao(entity.getDescricao());
 	}
 	public void delete(Long id) {
-		validationCategoriaId(id);
+		if(livroRepository.existsByCategoriasId(id)) throw new CategoriaDeletionException(id);
+		if(!repository.existsById(id)) throw new CategoriaNotFoundException(id);
 		repository.deleteById(id);
-	}
-	public void validationCategoriaId(Long id){
-		if(!repository.existsById(id)) {
-			throw new CategoriaNotFoundException(id);
-		}
-	}
-	public void badRequestId(Categoria categoria) {
-		if (categoria.getId() != null) {
-	        throw new BadRequestException();
-	    }
 	}
 }
